requirements.txt
langchain
pypdf2
python-dotenv
streamlit
tiktoken
faiss-cpu
openai
huggingface_hub
InstructorEmbedding
sentence_transformers